added multithreading for API
