+++ src/main/java/TelegramBotApp.java
@@ -1,14 +1,14 @@
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-import ua.goit.telegrambot.api.currency.FileUpdate;
+import ua.goit.telegrambot.api.currency.CurrencyUpdate;
 import ua.goit.telegrambot.telegram.TelegramCurrencyBot;
 
 public class TelegramBotApp {
     public static void main(String[] args) {
         //take and update currency exchange
-        FileUpdate fileUpdate = new FileUpdate();
-        fileUpdate.start();
+        CurrencyUpdate fileCurrencyUpdate = new CurrencyUpdate();
+        fileCurrencyUpdate.run();
 
         try {
             TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
