+++ src/main/resources/Currency_Mono_rates.json
@@ -1,0 +1,10 @@
+[
+{
+  "bank": "Mono",
+  "buyUSD": "29.2549991607666",
+  "saleUSD": "30.00029945373535",
+  "buyEUR": "31.729999542236328",
+  "saleEUR": "32.59980010986328",
+  "CrossCurseGBP": "39.455101013183594"
+}
+]
