+++ src/main/resources/Currency_NBU_rates.json
@@ -1,0 +1,8 @@
+[
+{
+  "bank": "NBU",
+  "rateUSD": "29.254899978637695",
+  "rateEUR": "31.82349967956543",
+  "rateGBP": "38.14250183105469"
+}
+]
