+++ src/main/resources/Currency_Privat_rates.json
@@ -1,0 +1,9 @@
+[
+{
+  "bank": "Privat",
+  "buyUSD": "29.254899978637695",
+  "saleUSD": "32.18000030517578",
+  "buyEUR": "31.795700073242188",
+  "saleEUR": "34.75"
+}
+]
