--- src/main/resources/Currency_rates.json
@@ -1,23 +1,0 @@
-[
-{
-  "bank": "NBU",
-  "rateUSD": "29.254899978637695",
-  "rateEUR": "31.795700073242188",
-  "rateGBP": "38.3765983581543"
-},
-{
-  "bank": "Privat",
-  "buyUSD": "29.254899978637695",
-  "saleUSD": "29.547399520874023",
-  "buyEUR": "31.579999923706055",
-  "saleEUR": "32.380001068115234"
-},
-{
-  "bank": "Mono",
-  "buyUSD": "29.2549991607666",
-  "saleUSD": "29.990400314331055",
-  "buyEUR": "31.899999618530273",
-  "saleEUR": "32.73970031738281",
-  "CrossCurseGBP": "39.218101501464844"
-}
-]
